tidying the process

+++ domain/src/main/scala/model/GameDetails.scala
@@ -1,3 +1,3 @@
 package model
 
-case class GameDetails(id: Long, gameId: String)
+case class GameDetails(id: Int, gameId: String)
